-add input and filter
