+++ package.json
@@ -15,6 +15,7 @@
     "@types/react-redux": "^7.1.21",
     "@types/styled-components": "^5.1.19",
     "axios": "^0.24.0",
+    "babel-plugin-styled-components": "^2.0.2",
     "google-map-react": "^2.1.10",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
@@ -47,8 +48,5 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
-  },
-  "devDependencies": {
-    "babel-plugin-styled-components": "^2.0.2"
   }
 }
