+++ src/store/reducers/SummarySlice.ts
@@ -9,7 +9,8 @@ interface ISummaryState {
     isLoading: boolean
     error: string
     show: boolean
-    data: string
+    date: string
+    searchCountry: string
 }
 
 const initialState: ISummaryState = {
@@ -22,7 +23,8 @@ const initialState: ISummaryState = {
         TotalRecovered: 0
     },
     countries: [],
-    data: '',
+    searchCountry: '',
+    date: '',
     isLoading: true,
     error: '',
     show: false
@@ -32,9 +34,9 @@ export const summarySlice = createSlice({
     name: 'summaryData',
     initialState,
     reducers: {
-       /*addCoords(state, action) {
-        state.countries[action.payload.]
-       }*/
+       setCountryName(state, action: PayloadAction<string>) {
+        state.searchCountry = action.payload
+       }
     },
     extraReducers: {
         [fetchSummaryData.fulfilled.type]: (state, action: PayloadAction<ISummaryDataResponse>) => {
@@ -42,7 +44,7 @@ export const summarySlice = createSlice({
             state.error = ''
             state.globalData = action.payload.Global
             state.countries = action.payload.Countries
-            state.data= action.payload.Date
+            state.date= action.payload.Date
         },
         [fetchSummaryData.pending.type]: (state) => {
             state.isLoading = true
@@ -54,4 +56,5 @@ export const summarySlice = createSlice({
     }
 })
 
+export const {setCountryName} = summarySlice.actions
 export default summarySlice.reducer
